feat: tarefa 2 finalizada

+++ src/services/category.services.ts
@@ -1,13 +1,20 @@
 import { prisma } from "../database/prisma";
+import { AppError } from "../middlewares/App.Error";
 import { TCategoryCreate, TCategoryReturn } from "../schemas/category.schema";
 
 export class CategoryServices{
-   async create(body: TCategoryCreate): Promise<TCategoryReturn>{
-      const newCategory = await prisma.category.create({data: body})
+   async create(userId: number, body: TCategoryCreate): Promise<TCategoryReturn>{
+      const categoryData = {...body, userId}
+      const newCategory = await prisma.category.create({data: categoryData})
       return newCategory
    }
    
-   async delete(id: number): Promise<void>{
+   async delete(userId: number, id: number): Promise<void>{
+      const category = await prisma.category.findUnique({where: {id}})
+      if(!category || category.userId !== userId){
+         throw new AppError("This user is not the category owner", 403)
+      }
+      
       await prisma.category.delete({where: {id}})
   }
 }
